Move temp stack vector into internal header and reduce including
